Use Typography in WorkspaceDetails
