+++ context/app/static/js/components/workspaces/WorkspaceDetails/index.js
@@ -1,0 +1,3 @@
+import WorkspaceDetails from './WorkspaceDetails';
+
+export default WorkspaceDetails;
