+++ context/app/static/js/components/workspaces/WorkspaceDetails/style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const Flex = styled.div`
+  display: flex;
+`;
+
+export { Flex };
